Update README.md
The smart menu ordering system in which the orders will be placed paperless using Arduino UNO, TFT touch display connected to the kitchen chef directly using ESP WIFI connectivity. This acted as the transmitting side. At the receiving side, the order placed is seen on Monitor/Screen through web server. Hence, the workers focused on billing and other work process and provided better quality services. We found out that this method is useful in many ways like- It not only saved the wastage of paper but also reduced the burden on the waiters and workers to great extent.